testing

+++ vm_riskxvii.c
@@ -286,7 +286,7 @@ int main(int argc, char **argv) {
 					reg[rd] = reg[rs1] & reg[rs2];
 				// SHIFT LEFT
 				} else if (func3 == 0b001 && func7 == 0) {
-					reg[rd] = reg[rs1] << reg[rs2];
+					reg[rd] = ((unsigned int) reg[rs1] << (unsigned int) reg[rs2];
 				// SHIFT RIGHT
 				} else if (func3 == 0b101 && func7 == 0) {
 					reg[rd] = ((unsigned int) reg[rs1] >> (unsigned int) reg[rs2]);
